Extract FollowsFilterMenuItems
- remove now unnecessary FilterPostsMenuItems
